+++ CHANGELOG.md
@@ -8,6 +8,8 @@ n2：💛 影响之前版本使用方式的更新（需要用户适配）
 n3：💚 不影响之前版本使用方式的更新（不需要用户适配）
 ```
 
+- 💚 new component for displaying progress：`Progress`
+
 ### NavBar
 
 - 💚 new props `hitSlop`：defines how far a touch event can start away from buttons in left and right